Refactor conditional selection
Now the cond_selector gets the 4 Tewds extended
coords of a point and a bit.
Verifies that the bit is either `1` or `0` and
then outputs the identity point if the bit was `0`
and the point itself if the bit was `1`.

This follows the idea mentioned in #8.

See https://github.com/dusk-network/bulletproofs_gadgets/pull/8#issuecomment-583831447